+++ products/AbstractBuilder.node.js
@@ -1,8 +1,12 @@
+//onsave jtree build produce AbstractBuilder.node.js
+const fs = require("fs")
 const { exec, execSync } = require("child_process")
 const recursiveReadSync = require("recursive-readdir-sync")
+const express = require("express")
 const { jtree } = require("../index.js")
 const { TypeScriptRewriter } = require("../products/TypeScriptRewriter.js")
 const { Disk } = require("../products/Disk.node.js")
+const registeredExtensions = { js: "//", maia: "doc.tooling ", ts: "//", grammar: "tooling ", gram: "tooling " }
 class AbstractBuilder extends jtree.TreeNode {
   _typeScriptToJavascript(sourceCode, forBrowser = false) {
     const ts = require("typescript")
@@ -54,6 +58,59 @@ class AbstractBuilder extends jtree.TreeNode {
   }
   _buildTsc(sourceCode, outputFilePath, forBrowser = false) {
     Disk.write(outputFilePath, this._typeScriptToJavascript(sourceCode, forBrowser))
+  }
+  _startServer(port, folder) {
+    const app = express()
+    app.listen(port, () => {
+      console.log(`Running builder. cmd+dblclick: http://localhost:${port}/`)
+    })
+    this._startListeningForFileChanges(folder)
+  }
+  _isProjectFile(path) {
+    return !path.includes("node_modules") && !path.includes("products") && !path.includes(".git")
+  }
+  _onFileChanged(fullPath) {
+    console.log(`SAVED ${fullPath}`)
+    if (!this._isProjectFile(fullPath)) {
+      console.log(` SKIP not a project file.`)
+      console.log(` DONE`)
+      return true
+    }
+    const fileExtension = jtree.Utils.getFileExtension(fullPath)
+    const commentPrefix = registeredExtensions[fileExtension]
+    if (!commentPrefix) {
+      console.log(` SKIP not a registered extension.`)
+      console.log(` DONE`)
+      return true
+    }
+    const prefix = commentPrefix + "onsave"
+    if (!Disk.exists(fullPath)) {
+      console.log(` SKIP file deleted`)
+      console.log(` DONE`)
+      return
+    } else {
+      const lines = Disk.read(fullPath).split("\n")
+      let line = lines.shift()
+      if (!line.startsWith(prefix)) console.log(` SKIP no tasks`)
+      while (line.startsWith(prefix)) {
+        const command = line.substr(prefix.length + 1)
+        const folder = jtree.Utils.getParentFolder(fullPath)
+        console.log(` COMMAND ${command}`)
+        const time = Date.now()
+        execSync(command, { cwd: folder, encoding: "utf8" })
+        console.log(` FINISHED ${Date.now() - time}ms`)
+        line = lines.shift()
+      }
+      console.log(` DONE`)
+    }
+  }
+  _startListeningForFileChanges(folder) {
+    const projectFolders = [folder] // todo
+    this._fsWatchers = projectFolders.map(folder =>
+      fs.watch(folder, { recursive: true }, (event, filename) => {
+        this._onFileChanged(folder + filename)
+      })
+    )
   }
   // todo: cleanup
   _getOutputFilePath(outputFileName) {