+++ products/jtable.node.js
@@ -1,4 +1,6 @@
 const { jtree } = require("../index.js")
+//onsave jtree build produce jtable.browser.js
+//onsave jtree build produce jtable.node.js
 // todo: create a Tree Language for number formatting
 const d3format = require("d3-format")
 const moment = require("moment")
@@ -844,6 +846,9 @@ class Column {
   getMax() {
     return this.getReductions().max
   }
+  getMin() {
+    return this.getReductions().min
+  }
   getMean() {
     return this.getReductions().mean
   }
@@ -1003,6 +1008,8 @@ const PrimitiveTypes = {
   HourMinute,
   CodeCol
 }
+//onsave jtree build produce jtable.browser.js
+//onsave jtree build produce jtable.node.js
 class Row {
   constructor(sourceObject = {}, table) {
     this._puid = this._getUniqueId()
@@ -1014,9 +1021,6 @@ class Row {
     return Row._uniqueId
   }
   destroy() {}
-  _getRowTable() {
-    return this._table
-  }
   async destroyRow() {}
   getAsArray(headerRow) {
     const obj = this.rowToObjectWithOnlyNativeJavascriptTypes()
@@ -1025,9 +1029,12 @@ class Row {
   getRowSourceObject() {
     return this._sourceObject
   }
+  toVector() {
+    return Object.values(this.rowToObjectWithOnlyNativeJavascriptTypes())
+  }
   // todo: rowToObjectWithOnlyNativeJavascriptTypes method? Its numerics where we need and strings where we need.
   _parseIntoObjectWithOnlyNativeJavascriptTypes() {
-    const columns = this._getRowTable().getColumnsMap()
+    const columns = this._table.getColumnsMap()
     const typedNode = {}
     Object.keys(columns).forEach(colName => {
       typedNode[colName] = this._getRowValueFromSourceColOrOriginalCol(colName)
@@ -1036,7 +1043,7 @@ class Row {
   }
   // why from source col? if we always copy, we shouldnt need that, correct? perhaps have an audit array of all operations on a row?
   _getRowValueFromSourceColOrOriginalCol(colName) {
-    const columns = this._getRowTable().getColumnsMap()
+    const columns = this._table.getColumnsMap()
     const destColumn = columns[colName]
     const sourceColName = destColumn.getSourceColumnName()
     const sourceCol = columns[sourceColName]
@@ -1080,6 +1087,8 @@ class Row {
   }
 }
 Row._uniqueId = 0
+//onsave jtree build produce jtable.browser.js
+//onsave jtree build produce jtable.node.js
 var TableParserIds
 ;(function(TableParserIds) {
   TableParserIds["csv"] = "csv"
@@ -2109,6 +2118,8 @@ And that has made all the difference.`
     [69, 80]
   ]
 }
+//onsave jtree build produce jtable.browser.js
+//onsave jtree build produce jtable.node.js
 class PivotTable {
   constructor(rows, inputColumns, outputColumns) {
     this._columns = {}
@@ -2292,6 +2303,19 @@ class Table {
   getJavascriptNativeTypedValues() {
     return this.getRows().map(row => row.rowToObjectWithOnlyNativeJavascriptTypes())
   }
+  toMatrix() {
+    return this.getRows().map(row => row.toVector())
+  }
+  toNumericMatrix() {
+    // todo: right now it drops them. should we 1 hot them?
+    const numericNames = this.getColumnsArray()
+      .filter(col => col.isNumeric())
+      .map(col => col.getColumnName())
+    return this.getRows().map(row => {
+      const obj = row.rowToObjectWithOnlyNativeJavascriptTypes()
+      return numericNames.map(name => obj[name])
+    })
+  }
   clone() {
     return new Table(this.cloneNativeJavascriptTypedRows())
   }
@@ -2425,7 +2449,9 @@ ${cols}
       .map(row => row.rowToObjectWithOnlyNativeJavascriptTypes())
       .map(obj => {
         Object.keys(nameMap).forEach(oldName => {
-          obj[nameMap[oldName]] = obj[oldName]
+          const newName = nameMap[oldName]
+          if (newName === oldName) return
+          obj[newName] = obj[oldName]
           delete obj[oldName]
         })
         return obj