+++ products/jtree.node.js
@@ -2442,10 +2442,14 @@ class TreeNode extends AbstractNode {
     this._getNodesByLineRegex(matches, regex)
     return matches
   }
+  // todo: remove?
   getNodesByLinePrefixes(columns) {
     const matches = []
     this._getNodesByLineRegex(matches, columns.map(str => new RegExp("^" + str)))
     return matches
+  }
+  nodesThatStartWith(prefix) {
+    return this.filter(node => node.getLine().startsWith(prefix))
   }
   _getNodesByLineRegex(matches, regs) {
     const rgs = regs.slice(0)
@@ -3105,7 +3109,7 @@ TreeNode.iris = `sepal_length,sepal_width,petal_length,petal_width,species
 4.9,2.5,4.5,1.7,virginica
 5.1,3.5,1.4,0.2,setosa
 5,3.4,1.5,0.2,setosa`
-TreeNode.getVersion = () => "49.4.0"
+TreeNode.getVersion = () => "49.5.0"
 class AbstractExtendibleTreeNode extends TreeNode {
   _getFromExtended(firstWordPath) {
     const hit = this._getNodeFromExtended(firstWordPath)